Requested PR changes
