+++ app/src/main/java/org/greenstand/android/TreeTracker/view/TextButton.kt
@@ -387,6 +387,16 @@ fun DepthButton(
     }
 }
 
+@Composable
+fun NoGPSDeviceDialog(onPositiveClick: () -> Unit){
+    CustomDialog(
+        dialogIcon = painterResource(id = R.drawable.error_outline),
+        title = stringResource(R.string.no_gps_device_header),
+        textContent = stringResource(R.string.no_gps_device_content),
+        onPositiveClick = onPositiveClick
+    )
+}
+
 enum class DepthSurfaceShape {
     Rectangle,
     Circle
